# tokenserv-api/src/main/java/com/anz/platform/common/tokenserv/exception/ErrorResponse.java
package com.anz.platform.common.tokenserv.exception;

public class ErrorResponse {

    private final String code;
    private final String message;

    public ErrorResponse(String code, String message) {
        this.code = code;
        this.message = message;
    }

    public String getCode() {
        return code;
    }

    public String getMessage() {
        return message;
    }
}