# tokenserv-infra/api-contract/swagger.yaml
openapi: 3.0.3
info:
  title: Tokenization Service API
  description: |
    A simple tokenization service that replaces sensitive account numbers
    with randomized tokens and allows reversing the process via detokenization.

    This API is intentionally minimal and designed for demonstration and
    interview evaluation purposes.
  version: 1.0.0
  contact:
    name: Diwakar
    email: diwakar.david@gmail.com

servers:
  - url: http://localhost:8080/tokenserv/api/v1/
    description: Local development server

components:
  schemas:
    TokenizerRequest:
      type: object
      properties:
        entity:
          type: array
          items:
            type: string
            minItems: 1
            maxItems: 100
          example:
            - "4111-1111-1111-1111"
            - "4444-3333-2222-1111"
            - "4444-1111-2222-3333"
      required:
        - entity

    TokenizerResponse:
      type: object
      properties:
        entity:
          type: array
          items:
            type: string
            minItems: 1
            maxItems: 100
          example:
            - "fvMymE7X0Je1IzMDgWooV5iGBPw0yoFy1"
            - "fvMymE7X0Je1IzMDgWooV5iGBPw0yoFy2"
            - "fvMymE7X0Je1IzMDgWooV5iGBPw0yoFy3"

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "TOKEN_NOT_FOUND"
        message:
          type: string
          example: "No entry found for encryption"

paths:
  /tokenize:
    post:
      summary: Tokenize account numbers
      description: |
        Accepts a list of account numbers and returns a list of generated tokens.
        Tokenization is idempotent: the same account number will always return
        the same token while the service is running.
      operationId: tokenizeAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenizerRequest'
      responses:
        "200":
          description: Successfully tokenized account numbers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenizerResponse'
        "400":
          description: Validation error (empty input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /detokenize:
    post:
      summary: Detokenize tokens
      description: |
        Accepts a list of tokens and returns the original account numbers.
        An error is returned if a token is unknown.
      operationId: detokenizeTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenizerRequest'
      responses:
        "200":
          description: Successfully detokenized tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenizerResponse'
        "400":
          description: Validation error (empty input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'